requirements.txt
tensorflow==2.15.0
keras==3.0.5
protobuf==6.31.0
numpy==1.24.4
matplotlib==3.7.1
pillow==10.0.0
flask==2.3.3
streamlit==1.28.2